SEGMENTATION; cpu.h functions as macros, fixed mmap_entry_t structure, posix-compliant putchar, basic support for 64-bit variadic args

+++ src/raw_video.h
@@ -40,7 +40,7 @@ int rvid_clrscr();
 
 int rvid_getpos(int* x, int* y);
 int rvid_setpos(int x, int y);
-int rvid_putchar(u8 c);
+int rvid_putchar(int c);
 int rvid_puts(const char* s);
 
 int rvid_vprintf(const char* f, va_list);
